Fix some bugs, and implement the main code.
Create the application code (main).

Change variable size (uint8_t to uint16_t).

Add post-increment to the data pointer.

+++ adxl345_code/include/adxl345.h
@@ -46,7 +46,7 @@ extern "C"{
 
 void ADXL345_init(SpiChannel_t Channel, DioPort_t Port, DioPin_t Pin);
 void ADXL345_read(SpiChannel_t Channel, DioPort_t Port, DioPin_t Pin, 
-uint8_t address, uint8_t *data);
+uint16_t address, uint16_t *data);
 
 #ifdef __cplusplus
 }   /*Extern C*/
